feat: initialize stocks tab module
